Attempted to make two tab animations asynchronous
Made the speed always 1 and modulus the total time to make sure the next frame should be played.

Also added a cast to header. (For some reason it was on footer but not header. Not sure whether or not to keep or remove the cast so I am keeping it to be safe).

Multitask is also made generic if speed is, for some reason, changed, it will still work (or should, might not if odd and even numbers come into play. Perhaps could use the lowest common factor of the two animation sequences? Could save performance, but not sure how to get that easily.)